reorganization of storage
