+++ hub/marray/array.py
@@ -1,66 +1,184 @@
-import warnings
-from hub.utils.store_control import StoreControlClient
-from hub.marray.interface import HubArray
+from pathos.threading import ThreadPool
 import numpy as np
-from hub.exceptions import WrongTypeError
+from hub.log import logger
+from .bbox import Bbox, chunknames, shade, Vec, generate_chunks
+from hub.backend.storage import Storage, S3, FS
+from hub.exceptions import IncompatibleBroadcasting, IncompatibleTypes, IncompatibleShapes, ArrayNotFound
+from hub.utils import StoreControlClient
+import json
 
 
-def _get_path(name, public=False):
-    if len(name.split('/')) == 1:
-        name = '{}/{}'.format(name, name)
-    user = name.split('/')[0]
-    dataset = name.split('/')[1].split(':')[0]
-    tag = name.split(':')
-    if len(tag) == 1:
-        tag.append('latest')
-    tag = tag[1]
-    bucket = StoreControlClient().get_config(public)['BUCKET']
-    if bucket == '':
-        exit()
-    path = 's3://'+bucket+'/'+user+'/'+dataset+'/'+tag
-    return path
+class HubArray(object):
+    def __init__(self, shape=None, chunk_shape=None, dtype=None, key=None, protocol='file', parallel=True, order='F', public=False, storage=None):
+        self.shape = shape
+        self.chunk_shape = chunk_shape
+        self.dtype = dtype
+        self.key = key
+        self.protocol = protocol
+        self.storage = storage if storage is not None else S3(StoreControlClient.get_config(public)['BUCKET'], public=public)
+        self.order = order
 
+        parallel = 25 if parallel else 1
 
-def load(name):
-    is_public = name in ['imagenet', 'cifar', 'coco', 'mnist']
-    path = _get_path(name, is_public)
-    return HubArray(key=path, public=is_public)
+        self.pool = ThreadPool(nodes=parallel)
+        self.initialize(self.key)
 
+        # Make sure the object was properly initialized
+        assert self.shape
+        assert self.chunk_shape
+        assert self.dtype
 
-def array(shape=None, name=None, dtype='float', chunk_size=None):
+    def initialize(self, path):
+        cloudpath = "{}/info.txt".format(path)
+        info = self.storage.get(cloudpath)
 
-    if not name:
-        raise Exception(
-            'No name provided, please name your array - hub.array(..., name="username/dataset:version") ')
+        if not info and not self.shape:
+            name, dataset, version = self.key.split('/')[-3:]
+            raise ArrayNotFound(
+                'Could not identify array with name {}/{}:{}. Please make sure the array name is correct.'.format(name, dataset, version))
+        if info:
+            info = json.loads(info.decode('utf-8'))
+            self.shape = info['shape']
+            self.chunk_shape = info['chunk_shape']
+            self.key = info['key']
+            self.dtype = info['dtype']
+            self.order = info['order']
+            self.protocol = info['protocol']
+        else:
+            info = {
+                'shape': self.shape,
+                'chunk_shape': self.chunk_shape,
+                'key': self.key,
+                'dtype': self.dtype,
+                'order': self.order,
+                'protocol': self.protocol,
+            }
+            info = json.dumps(info).encode('utf-8')
+            self.storage.put(cloudpath, info, content_type=None)
 
-    path = _get_path(name)
+    def generate_cloudpaths(self, slices):
+        # Slices -> Bbox
+        slices = Bbox(Vec.zeros(self.shape), self.shape).reify_slices(slices)
+        requested_bbox = Bbox.from_slices(slices)
 
-    if not shape:
-        return load(name)
+        # Make sure chunks fit
+        full_bbox = requested_bbox.expand_to_chunk_size(
+            self.chunk_shape, offset=Vec.zeros(self.shape)
+        )
 
-    return create(path, shape, dtype, chunk_size)
+        # Clamb the border
+        full_bbox = Bbox.clamp(full_bbox, Bbox(
+            Vec.zeros(self.shape), self.shape))
 
+        # Generate chunknames
+        cloudpaths = list(chunknames(
+            full_bbox, self.shape,
+            self.key, self.chunk_shape,
+            protocol=self.protocol
+        ))
 
-def create(path, dim=[50000, 28, 28], dtype='float', chunk_size=None):
-    # auto chunking
-    if chunk_size is None:
-        chunk_size = list(dim)
-        chunk_size[0] = 1
+        return cloudpaths, requested_bbox
 
-    # Input checking
-    assert len(chunk_size) == len(dim)
-    assert np.array(dim).dtype in np.sctypes['int']
-    assert np.array(chunk_size).dtype in np.sctypes['int']
+    # read from raw file and transform to numpy array
+    def decode(self, chunk):
+        return np.frombuffer(bytearray(chunk), dtype=self.dtype).reshape(self.chunk_shape, order='F')
 
-    try:
-        dtype = np.dtype(dtype).name
-    except:
-        raise WrongTypeError('Dtype {} is not supported '.format(dtype))
+    def download_chunk(self, cloudpath):
+        chunk = self.storage.get(cloudpath)
+        if chunk:
+            chunk = self.decode(chunk)
+        else:
+            chunk = np.zeros(shape=self.chunk_shape,
+                             dtype=self.dtype, order=self.order)
+        bbox = Bbox.from_filename(cloudpath)
+        return chunk, bbox
 
-    return HubArray(shape=dim, dtype=dtype, chunk_shape=chunk_size, key=path, protocol=None)
+    def download(self, cloudpaths, requested_bbox):
+        # Download chunks
+        chunks_bboxs = self.pool.map(self.download_chunk, cloudpaths)
 
+        # Combine Chunks
+        renderbuffer = np.zeros(
+            shape=requested_bbox.to_shape(), dtype=self.dtype, order=self.order)
 
-def delete(name):
-    path = _get_path(name)
-    bucket = StoreControlClient().get_config()['BUCKET']
-    s3.Object(bucket, path.split(bucket)[-1]).delete()
+        def process(chunk_bbox):
+            chunk, bbox = chunk_bbox
+            shade(renderbuffer, requested_bbox, chunk, bbox)
+        self.pool.map(process, chunks_bboxs)
+
+        return renderbuffer
+
+    def squeeze(self, slices, tensor):
+        squeeze_dims = []
+
+        if isinstance(slices, list) and len(slices) == 1:
+            slices = slices[0]
+
+        if not isinstance(slices, list) and not isinstance(slices, tuple):
+            slices = [slices]
+
+        for dim in range(len(slices)):
+            if isinstance(slices[dim], int):
+                squeeze_dims.append(dim)
+
+        if len(squeeze_dims) >= 1:
+            tensor = tensor.squeeze(axis=(*squeeze_dims, ))
+
+        if len(tensor.shape) == 0:
+            tensor = tensor.item()
+
+        return tensor
+
+    def __getitem__(self, slices):
+        cloudpaths, requested_bbox = self.generate_cloudpaths(slices)
+        tensor = self.download(cloudpaths, requested_bbox)
+        tensor = self.squeeze(slices, tensor)
+        return tensor
+
+    def encode(self, chunk):
+        return chunk.tostring('F')
+
+    def upload_chunk(self, cloudpath_chunk):
+        cloudpath, chunk = cloudpath_chunk
+        chunk = self.encode(chunk)
+        chunk = self.storage.put(cloudpath, chunk, content_type=None)
+
+    def chunkify(self, cloudpaths, requested_bbox, item):
+        chunks = []
+        for path in cloudpaths:
+            cloudchunk = Bbox.from_filename(path)
+            intersection = Bbox.intersection(cloudchunk, requested_bbox)
+            chunk_slices = (intersection-cloudchunk.minpt).to_slices()
+            item_slices = (intersection-requested_bbox.minpt).to_slices()
+
+            chunk = np.zeros(shape=self.chunk_shape,
+                             dtype=self.dtype, order=self.order)
+            if np.any(np.array(intersection.to_shape()) != np.array(self.chunk_shape)):
+                logger.debug('Non aligned write')
+                chunk, _ = self.download_chunk(path)
+            else:
+                chunk = np.zeros(shape=self.chunk_shape,
+                                 dtype=self.dtype, order=self.order)
+
+            chunk[chunk_slices] = item[item_slices]
+            chunks.append(chunk)
+
+        return zip(cloudpaths, chunks)
+
+    def upload(self, cloudpaths, requested_bbox, item):
+        try:
+            item = np.broadcast_to(item, requested_bbox.to_shape())
+        except ValueError as err:
+            raise IncompatibleBroadcasting(err)
+
+        try:
+            item = item.astype(self.dtype)
+        except Exception as err:
+            raise IncompatibleTypes(err)
+
+        cloudpaths_chunks = self.chunkify(cloudpaths, requested_bbox, item)
+        self.pool.map(self.upload_chunk, list(cloudpaths_chunks))
+
+    def __setitem__(self, slices, item):
+        cloudpaths, requested_bbox = self.generate_cloudpaths(slices)
+        self.upload(cloudpaths, requested_bbox, item)
